refactor(toolkit): move paint primitives down from app_skeleton
The desktop shell depends on the toolkit but not on app_skeleton, so the
primitive layer was unreachable from the chrome that needs it most. Moving
it down gives both callers one implementation instead of a second copy in
the shell's own fill.rs. app_skeleton re-exports the module, so its
dependents compile unchanged.
