+++ userland/toolkit/src/paint/clear.rs
