+++ userland/toolkit/src/paint/fill_rect.rs
