+++ userland/toolkit/src/paint/round_stroke.rs
