+++ userland/toolkit/src/paint/text_ttf.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use nonos_toolkit::font::ttf;
+use crate::font::ttf;
 
 use super::buffer::PaintBuffer;
